+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - feat: add kortex API support for aarch64 (#1)
 - fix(build): update kortex api links (#2)
 - feat(description): add urdf example for gen3 7dof (#3)
+- feat(build): add package toml and script to build collection (#4)